\name{metaBLUEX-package}
\alias{metaBLUEX-package}
\alias{metaBLUEX}
\docType{package}
\title{
\packageTitle{metaBLUEX}
}
\description{
\packageDescription{metaBLUEX}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{metaBLUEX}
\packageIndices{metaBLUEX}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{metaBLUEX}

Maintainer: \packageMaintainer{metaBLUEX}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
