% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaBLUEX.R
\name{metabluex}
\alias{metabluex}
\alias{print.metabluex}
\title{Magic Wrapper for 'BLUE Estimators for Mean and Standard Deviation'}
\usage{
metabluex(summary = NULL, n = NULL, scenario = "S1")

\method{print}{metabluex}(object, ...)
}
\arguments{
\item{summary}{a vector of ordered summary statistics in ascending order.}

\item{n}{the sample size.}

\item{scenario}{a character string indicating the scenario in which summary statistics is reported. The options for the \code{scenario} argument are as follows.
\tabular{ll}{
\code{"S1"} \tab for the median, minimum and maximum values. \cr
\code{"S2"} \tab for the median, first and third quartiles. \cr
\code{"S3"} \tab for the median, first and third quartiles, and minimum and maximum values. \cr
\code{"tertiles"} \tab for tertiles. \cr
\code{"quintiles"} \tab for quintiles. \cr
\code{"deciles"} \tab for deciles. \cr}}

\item{object}{an object of class \code{metabluex}.}

\item{...}{additional arguments (not used).}
}
\value{
A list with class "metabluex" containing the following components:
\tabular{ll}{
\code{summary} \tab summary statistics. \cr
\code{n} \tab the sample size. \cr
\code{scenario} \tab the scenario in which summary statistics is reported. \cr
\code{result} \tab a list containing results (\eqn{\hat\mu} and \eqn{\hat\sigma} as a vector) in all distributions and methods applicable for the selected scenario. \cr }
}
\description{
A magic wrapper for \code{metablue} to provide a colorful and detailed report for applicable distributions and methods in estimation, once and for all.
}
\details{
\code{metabluex} is a magic wrapper for \code{metablue}, which outputs all possible estimations in all distributions and methods applicable for the selected scenario.

This function is highly useful in scenarios where you have determined summary statistics and want to explore different methods and distributional assumptions for estimating the mean and standard deviation.
You may acquire a tidy report to elegantly review and compare all possible BLUE methods applicable under the current scenario.

Consistent with the applicability of \code{"dist"} and \code{"method"} in \code{metablue},
you only need to specify \code{"scenario"} to get the estimation results.
Generally, you don’t need to check about the details below:
\itemize{
   \item if \code{scenario} is \code{"S1"}, \code{"S2"} or \code{"S3"}, report contains estimates in \code{"normal"} for \code{"wls"}, \code{"bala"}, \code{"yang"}, in \code{"laplace"} for \code{"yang"}, \code{"approx"} and in \code{"logistic"} for \code{"yang"}, \code{"approx"}.
   \item if \code{scenario} is \code{"tertiles"}, \code{"quintiles"} or \code{"deciles"}, report contains estimates in \code{"normal"} for \code{"bala"}, \code{"yang"}, in \code{"laplace"} for \code{"yang"} and in \code{"logistic"} for \code{"yang"}.
}
\code{crayon} package is used to report in a colorful way.
}
\examples{
set.seed(1)
n = 100
# mu = 1, sigma = 1
r = rnorm(n, 1, 1)

### S3 ###
fivenumber = fivenum(r)
metabluex(fivenumber, n, "S3")

### tertiles ###
tertiles = quantile(r, c(1/3, 2/3), names = FALSE)
metabluex(tertiles, n, "tertiles")

}
\author{
Sung Manhin \email{songwenxuan@ruc.edu.cn}
}
