% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaBLUEX.R
\name{metablue}
\alias{metablue}
\title{BLUE Estimators for Mean and Standard Deviation}
\usage{
metablue(
  summary = NULL,
  n = NULL,
  scenario = "S1",
  dist = "normal",
  method = "wls"
)
}
\arguments{
\item{summary}{a vector of ordered summary statistics in ascending order.}

\item{n}{the sample size.}

\item{scenario}{a character string indicating which scenario of summary statistics is reported. The options for the \code{scenario} argument are as follows. 
\tabular{ll}{
\code{"S1"} \tab for the median, minimum and maximum values. \cr
\code{"S2"} \tab for the median, first and third quartiles. \cr
\code{"S3"} \tab for the median, first and third quartiles, and minimum and maximum values. \cr
\code{"tertiles"} \tab for tertiles. \cr
\code{"quintiles"} \tab for quintiles. \cr
\code{"deciles"} \tab for deciles. \cr}}

\item{dist}{a character string indicating which distribution of data is assumed. The options for the \code{dist} argument are as follows.
\tabular{ll}{
\code{"normal"} \tab for the normal distribution. \cr
\code{"laplace"} \tab for the Laplace distribution. \cr
\code{"logistic"} \tab for the logistic distribution. \cr}}

\item{method}{a character string indicating which method to use. The options for the \code{method} argument are as follows.
\tabular{ll}{
\code{"wls"} \tab an abbreviation for approximation formulae of BLUE estimators proposed by Wan et al. (2014), Luo et al. (2018) and Shi et al. (2020). Only applicable when \code{dist} is set to \code{"normal"} and \code{scenario} is set to \code{"S1"}, \code{"S2"} or \code{"S3"}. \cr
\code{"bala"} \tab Balakrishnan et al. (2022)'s approximation method. Only applicable when \code{dist} is set to \code{"normal"}. \cr
\code{"yang"} \tab Yang et al. (2022)'s approximation method. \cr
\code{"approx"} \tab proposed by the authors of this package to handle the Laplace and logistic distribution. \cr}}
}
\value{
A vector of length two, the first element is \eqn{\hat\mu} and the second one is \eqn{\hat\sigma}.
}
\description{
A comprehensive function for estimating the mean and standard deviation from summary statistics, incorporating both published and novel BLUE methods.
}
\details{
\code{metablue} implements BLUE methods for estimating the mean and standard deviation from summary statistics, 
including methods proposed by the authors. 
Below are important notes regarding the function's arguments.
\itemize{
   \item \code{method = "wls"} can only be used under normality assumption and three basic scenarios. 
   \item \code{method = "bala"} can only be used under normality assumption. Balakrishnan et al. (2022)'s approximation method is proposed for three basic scenarios only, and the authors of this package extend their method to tertiles, quintiles and deciles. Additional message will be printed when using the method in extended scenarios. 
   \item \code{method = "yang"} can be used for all six scenarios under normality assumption. The authors of this package extend their method to the Laplace and logistic distribution. Additional message will be printed when using the method in extended distributions.
   \item \code{method = "approx"} is proposed for the Laplace and logistic distribution and three basic scenarios, inspired by Wan et al. (2014), Luo et al. (2018) and Shi et al. (2020)'s works.
}
}
\examples{
set.seed(1)
n=100
r=rnorm(n, 1, 1)
fivenumber=fivenum(r)
metablue(fivenumber, n, "S3")

}
\references{
Luo D, Wan X, Liu J, and Tong T. (2016). Optimally estimating the sample mean from the sample size, median, mid-range, and/or mid-quartile range. \emph{Statistical Methods in Medical Research}, arXiv:1505.05687.
}
\author{
Sung Manhin \email{songwenxuan@ruc.edu.cn}
}
